auth-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
aiomysql==0.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
alembic==1.13.1
email-validator==2.1.0
